feat!: Unify include types
This deduplicates types where includes are possible on API responses
with those that are the same resource but no includes come back.

Includes are optional on the type and thus can be used interchangeably

BREAKING CHANGE: Include types are renamed to replace their
non-include variant, see UPGRADING.md for migration details.